requirements.txt
flask
scikit-learn
python-dotenv
gunicorn
cohere